+++ framework/node/packages/acm-sdk/src/index.ts
@@ -10,3 +10,4 @@ export * from './nucleus.js';
 export * from './context.js';
 export * from './context-provider.js';
 export * from './utils.js';
+export { universalDigest } from './hash.js';